+++ doc/src/history.md
@@ -5,7 +5,9 @@
 ### Fixes
 
 - `om ci`
-  - Allow `--override-input` to work again (#439)
+  - Extra nix handling
+      - Allow `--override-input` to work again (#439)
+      - Support `--rebuild` by disallowing it in irrelevant subcommands (`eval`, `flake lock`)
 
 ## 1.0.2 (2025-03-11) {#1.0.2}